requirements.txt
streamlit==1.40.0
google-adk==1.1.1
opik==1.7.28
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
